+++ src/layouts/Components/Footer/index.js
@@ -1,0 +1,3 @@
+import Copyright from './Footer';
+
+export default Copyright;
