+++ src/layouts/Components/index.js
@@ -1,0 +1,4 @@
+import ButtonAppBar from './NavBar';
+import Copyright from './Footer';
+
+export { ButtonAppBar, Copyright };
